+++ src/components/Listener.tsx
@@ -1,73 +1,36 @@
 import React, { ReactNode } from 'react'
 import { app, action } from "../imports"
-import { ActionDescriptor} from "photoshop/dist/types/photoshop"
-import { Descriptor } from 'photoshop/dist/types/UXP'
+import { ActionDescriptor, BasicActionDescriptor} from "photoshop/dist/types/photoshop"
+import { IAppState, IAction, ISettings } from '../reducers/initialState'
+import ActionItemContainer from './ActionItemContainer'
+import Filter from './Filter'
+import FilterContainer from './FilterContainer'
 
-interface AppState {
-	listening: boolean
-	currentID: number
-	collapsed:boolean
+export interface IListenerProps{
+	settings: ISettings
 	actions: IAction[]
-	batchPlayDecorator: boolean
-	lastHistoryID:number
 }
 
-interface IAction{
-	id:number
-	title:string
-	collapsed:boolean
-	descriptor: ActionDescriptor
-	timeCreated: number
-	playReplies: IPlayReply[]
+export interface IListenerDispatch{
+	setCollapsed(enabled: boolean): void
+	setListener(enabled:boolean):void
+	addAction(action:IAction):void
+	setBatchPlayDecorator(enabled:boolean):void
+	clearLog():void
+	setLastHistoryID(id: number): void
+	incrementActionID():void
 }
 
-interface IPlayReply{
-	descriptors:Descriptor[]
-	time:number
-}
+type TListener = IListenerProps & IListenerDispatch
+
 
 /**
  * Usage of the Photoshop event listening APIs.
  * This section of the APIs are subject to sweeping changes.
  */
-export default class ListenerComponent extends React.Component<{}, AppState> {
-	constructor(props) {
+export default class Listener extends React.Component<TListener> {
+	constructor(props:TListener) {
 		super(props)
-		this.state = {
-			currentID: 0,
-			collapsed:true,
-			batchPlayDecorator:false,
-			listening: false,
-			actions: [],
-			lastHistoryID:-1
-		}
-
-		this.listener = this.listener.bind(this)
-	}
-
-	private decorateCode(desc: ActionDescriptor): string{
-		const codeStr = JSON.stringify(desc, null, 3)
-
-		if (this.state.batchPlayDecorator) {
-			const result = "await PhotoshopAction.batchPlay([\n"+codeStr+"\n], {})";
-			return result;			
-		} else {
-			return codeStr;
-		}
-	}
-
-	private toggleExpandAction(index:number) {
-		this.setState((prevState) => {
-			const obj = {
-				...prevState,
-				actions: [
-					...prevState.actions
-				]
-			}
-
-			obj.actions[index].collapsed = !obj.actions[index].collapsed;
-			return obj;
-		})
 	}
 
 	private async getHistoryState() {
@@ -84,49 +47,22 @@ export default class ListenerComponent extends React.Component<{}, AppState> {
 				]
 			}
 		], {});
-		console.log(test[0].name);
 
 		let result = ""
 
-		if (test[0].ID!==this.state.lastHistoryID && this.state.lastHistoryID !== -1) {
+		if (test[0].ID!==this.props.settings.lastHistoryID && this.props.settings.lastHistoryID !== -1) {
 			result =  " | "+test[0].name;
 		}
 
-		this.setState((prevState) => {
-			return {
-				...prevState,
-				lastHistoryID:test[0].ID
-			}
-		})
+		this.props.setLastHistoryID(test[0].ID);
 
 		return result;
 	}
 
-	private renderActions() {
-		const actions = this.state.actions;
+	private renderActions=()=> {
+		const actions = this.props.actions;
 		if (actions.length) {
-			const elements: ReactNode[] = [];
-		
-			for (let i = 0; i < actions.length; i++) {
-				const action = actions[i];
-				elements.unshift(
-					<div className="action" key={i}>
-						<div className="header">
-							<sp-button quiet variant="primary"  onClick={() => this.toggleExpandAction(i)} className="collapse">{action.collapsed ? "⮞" : "⮟"}{" " + action.title}</sp-button>
-							<div className="spread"></div>
-							<sp-button quiet variant="secondary" title="play" onClick={() => this.play(i)}>Play</sp-button>
-							<sp-button quiet variant="secondary" title="copy" onClick={() => this.copy(i)}  >Copy</sp-button>
-						</div>
-						{
-							action.collapsed ? null :
-								<div className="code">
-									{this.decorateCode(action.descriptor)}
-								</div>
-						}
-					</div>
-				)
-			}
-			
+			const elements: ReactNode[] = actions.map(action => <ActionItemContainer key={action.id} action={action}/>);
 			return elements;
 		} else {
 			return (
@@ -138,133 +74,75 @@ export default class ListenerComponent extends React.Component<{}, AppState> {
 	}
 
 	private clearLog() {
-		this.setState((prevState) => {
-			return {
-				...prevState,
-				actions: []
-			}
+		this.props.clearLog();
+	}
+
+	private toggleCollapseOption = (e:any) => {
+		const value = e.target.checked;
+		this.props.setCollapsed(value);
+	}
+
+	private onBatchPlay=(enabled:boolean)=> {		
+		this.props.setBatchPlayDecorator(enabled);
+	}
+
+	/**
+	 * Listener to be attached to all Photoshop notifications.
+	 */
+	public  listener = async (event: string, message: BasicActionDescriptor) => {
+		const {collapsed} = this.props.settings;
+
+		const historyName = await this.getHistoryState();
+		this.props.addAction({
+			collapsed:collapsed,
+			descriptor: {
+				_obj: event,
+				...message
+			},
+			id: NaN, // will be set within reducer
+			playReplies: [],
+			timeCreated: Date.now(),
+			title: event + historyName,
 		})
+		this.props.setLastHistoryID
+	}
+
+	/**
+	 * Attaches the simple listener to the app.
+	 */
+	private attachListener = async() => {
+		app.eventNotifier = this.listener		
+		this.props.setListener(true);
+	}
+
+
+	/**
+	 * Attaches a null listener to the app.
+	 */
+	private removeListener = async() => {
+		const app = require('photoshop').app
+		app.eventNotifier = undefined
+
+		this.props.setListener(false);
 	}
 
 	public render() {
-		const batchPlay = this.state.batchPlayDecorator;
-		const collapsed = this.state.collapsed;
+		const batchPlay = this.props.settings.batchPlayDecorator;
 
 		return (
 			<div className="component">
+				<FilterContainer />
 				<div className="actionList">
 					{this.renderActions()}
 				</div>
 				<div className="flex-row start controls">
-					{this.state.listening ? <sp-button variant="negative" onClick={() => this.removeListener()}>Stop</sp-button> : <sp-button variant="cta" onClick={() => this.attachListener()}>Start</sp-button>}
-					<sp-button quiet variant="secondary" onClick={() => this.clearLog()}>Clear</sp-button>
+					{this.props.settings.listening ? <sp-button variant="primary" onClick={this.removeListener}>Stop</sp-button> : <sp-button variant="cta" onClick={this.attachListener}>Start</sp-button>}
+					<sp-button quiet variant="secondary" onClick={this.clearLog}>Clear</sp-button>
 					{batchPlay ? <sp-button quiet variant="secondary" onClick={() => this.onBatchPlay(false)}>Raw code</sp-button > : <sp-button quiet variant="secondary" onClick={() => this.onBatchPlay(true)}>Batch play code</sp-button >}
 					<sp-checkbox onClick={this.toggleCollapseOption} checked>Collapsed</sp-checkbox>
 				</div>
 
 			</div>
 		);
-	}
-
-	private toggleCollapseOption = (e) => {
-		const value = e.target.checked;
-		console.log(value);
-		this.setState((prevState) => {			
-			return {
-				...prevState,
-				collapsed:value
-			}
-		})
-	}
-
-	private onBatchPlay(enabled:boolean) {
-		this.setState((prevState) => {
-			return {
-				...prevState,
-				batchPlayDecorator: enabled
-			}
-		})
-	}
-
-	/**
-	 * Listener to be attached to all Photoshop notifications.
-	 */
-	public async listener(event: string, message: ActionDescriptor) {
-		const historyName = await this.getHistoryState();
-		this.setState((prevState) => {
-			message = {
-				_obj: event,
-				...message
-			}
-
-			return {
-				actions: [
-					...prevState.actions,
-					{
-						collapsed:prevState.collapsed,
-						descriptor: message,
-						id: prevState.currentID,
-						title: message._obj + historyName,
-						timeCreated: Date.now(),
-						playReplies: []
-					}
-					
-				]
-			}
-		})
-	}
-
-	/**
-	 * Attaches the simple listener to the app.
-	 */
-	private async attachListener() {
-		app.eventNotifier = this.listener
-		this.setState({ listening: true })
-	}
-
-	private copy(index: number) {
-		const text = this.decorateCode(this.state.actions[index].descriptor);
-		this.copyToClipboard(text);
-	}
-
-	private async play(index: number) {
-		const result = await action.batchPlay([this.state.actions[index].descriptor], {});
-
-		this.setState((prevState) => {
-			const obj = {
-				...prevState,
-				actions: [
-					...prevState.actions
-				]
-			}
-
-			obj.actions[index].playReplies = [
-				...obj.actions[index].playReplies,
-				{
-					descriptors: result,
-					time: Date.now(),
-				}
-			]
-			return obj;
-		})
-	}
-
-	private async copyToClipboard(text:string) {
-		await action.batchPlay([{
-			_obj: "textToClipboard",
-			textData: text
-		}], {})
-	}
-
-	/**
-	 * Attaches a null listener to the app.
-	 */
-	private async removeListener() {
-		const app = require('photoshop').app
-		app.eventNotifier = undefined
-
-		this.setState({
-			listening: false,
-		})
 	}
 }