+++ CHANGELOG.md
@@ -1,12 +1,16 @@
 # Changelog
 
+## 0.4.0
+- feature: adds possibility to filter by name (case sensitive)
+- code: implements Redux store
+
 ## 0.3.0
 - feature: adds play button
 - improvements: UI
 - code: adds types
 
 ## 0.2.0
--  feature: collapsable actions
+- feature: collapsable actions
 - feature: default option for collapsable action
 
 ## 0.1.0